refactor: move dotfiles to resources/dotfiles
