+++ internal/orchestrator/reconcile.go
@@ -9,8 +9,8 @@ import (
 
 	"os"
 
-	"github.com/richhaase/plonk/internal/dotfiles"
 	"github.com/richhaase/plonk/internal/lock"
+	"github.com/richhaase/plonk/internal/resources/dotfiles"
 	"github.com/richhaase/plonk/internal/resources/packages"
 	"github.com/richhaase/plonk/internal/state"
 )
