+++ internal/orchestrator/sync.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 
 	"github.com/richhaase/plonk/internal/config"
-	"github.com/richhaase/plonk/internal/dotfiles"
+	"github.com/richhaase/plonk/internal/resources/dotfiles"
 	"github.com/richhaase/plonk/internal/resources/packages"
 	"github.com/richhaase/plonk/internal/state"
 )
