+++ internal/resources/dotfiles/atomic.go
