+++ internal/resources/dotfiles/operations_test.go
