+++ internal/resources/dotfiles/scanner.go
