+++ internal/resources/dotfiles/scanner_test.go
